Improve problem 662 solution
Instead of using the negative and positive range of an int to map the
node coordinates, simply use positive values. This makes the math much
simpler, as we just have to multiply the node position by 2 and for the
right node add 1.

Sadly had to switch everything for longs since Java does not have
unsigned ints lol.